pyproject.toml
[tool.poetry]
name = "scrapper"
version = "0.1.0"
description = ""
authors = ["warismstf <warismstf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
environs = "^11.0.0"
psycopg = "^3.2.1"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
channels = "^4.1.0"
redis = "^5.0.7"
celery = "^5.4.0"
django-celery-results = "^2.5.1"
channels-redis = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
